selenium_tests/requirements-dev.txt
selenium>=4.0
pytest>=7.0
python-dotenv>=1.0
pytest-html>=3.0
